🔥 Remove visitor-based API
