+++ crates/nodui_egui/src/lib.rs
@@ -2,18 +2,11 @@
 //!
 //! [nodui]: https://crates.io/crates/nodui
 
-pub mod connection;
-pub mod context_menu;
-mod conversion;
-mod editor;
 mod editor2;
 mod misc;
 mod socket;
 mod viewport;
-mod visitor;
 
-pub use connection::CustomConnectionRenderer;
-pub use editor::{GraphEditor, GraphOutput};
 pub use socket::RenderedSocket;
 pub use viewport::Viewport;
