+++ crates/nodui_egui/src/misc/mod.rs
@@ -1,5 +1,4 @@
 //! Miscellaneous
 
-pub(crate) mod collect;
 pub(crate) mod collector;
 pub(crate) mod layout;
