+++ examples/playground/src/app/adapter/mod.rs
@@ -1,21 +1,17 @@
-mod node;
-
 use std::ops::{Deref, DerefMut};
 
-use node::NodeAdapter;
-use nodui::{ConnectionHint, NodeSeq, SizeHint};
 use serde::{Deserialize, Serialize};
 
 use crate::graph;
 
 #[derive(Default, Serialize, Deserialize)]
-pub struct GraphAdapter {
+pub struct GraphApp {
     pub graph: graph::Graph,
     pub selected_node: Option<graph::NodeId>,
     pub clipboard: Option<(graph::NodeStyle, Vec<graph::SocketStyle>)>,
 }
 
-impl Deref for GraphAdapter {
+impl Deref for GraphApp {
     type Target = graph::Graph;
 
     fn deref(&self) -> &Self::Target {
@@ -23,41 +19,13 @@ impl Deref for GraphAdapter {
     }
 }
 
-impl DerefMut for GraphAdapter {
+impl DerefMut for GraphApp {
     fn deref_mut(&mut self) -> &mut Self::Target {
         &mut self.graph
     }
 }
 
-impl nodui::GraphAdapter for GraphAdapter {
-    type NodeId = graph::NodeId;
-    type SocketId = graph::SocketId;
-
-    fn accept<'graph, V>(&'graph mut self, mut visitor: V)
-    where
-        V: nodui::GraphVisitor<'graph, Self::NodeId, Self::SocketId>,
-    {
-        let graph::ViewMut { nodes, connections } = self.graph.view_mut();
-        let mut node_seq = visitor.nodes(SizeHint::of(nodes));
-        for node in nodes {
-            node_seq.visit_node(NodeAdapter { node, connections });
-        }
-    }
-
-    fn connection_hint(&self, _a: Self::SocketId, _b: Self::SocketId) -> ConnectionHint {
-        ConnectionHint::Accept // TODO
-    }
-
-    fn connect(&mut self, a: Self::SocketId, b: Self::SocketId) {
-        self.connections_mut().connect(a, b);
-    }
-
-    fn connections(&self) -> impl Iterator<Item = (Self::SocketId, Self::SocketId)> {
-        self.graph.connections().iter()
-    }
-}
-
-impl GraphAdapter {
+impl GraphApp {
     pub fn selected_node(&mut self) -> Option<&mut graph::Node> {
         self.selected_node
             .and_then(|node_id| self.graph.get_node_mut(node_id))
