--- playground/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "dev_playground"
-version = "0.0.0"
-edition = "2021"
-
-description = "A playground to test and debug the nodui crate"
-
-publish = false
-
-
-[dependencies]
-nodui = { workspace = true, features = ["serde"] }
-
-eframe = { workspace = true, features = ["persistence"] }
-
-egui.workspace = true
-
-serde = "1"
-uuid = { version = "1.8.0", default-features = false, features = [
-    "v4",
-    "serde",
-] }
