--- playground/src/main.rs
@@ -1,19 +1,0 @@
-mod app;
-mod graph;
-
-use app::App;
-
-fn main() -> eframe::Result<()> {
-    let native_options = eframe::NativeOptions {
-        viewport: egui::ViewportBuilder::default()
-            .with_inner_size([400.0, 300.0])
-            .with_min_inner_size([300.0, 220.0]),
-        ..Default::default()
-    };
-
-    eframe::run_native(
-        "Nodui Playground",
-        native_options,
-        Box::new(|cc| Ok(Box::new(App::new(cc)))),
-    )
-}
